+++ src/Http/Controllers/ChronicleUiController.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Chronicle\Http\Controllers;
 
 use Chronicle\Checkpoints\Checkpoint;
-use Chronicle\Entry\Entry;
 use Chronicle\Facades\Chronicle;
 use Chronicle\Query\LedgerStats;
 use Illuminate\Http\RedirectResponse;